+++ cookbook/seismic_epic2d_iterate.py
@@ -1,5 +1,5 @@
 """
-Seis: Show steps taken by different algorithms for 2D epicenter estimation
+seismic: Show steps taken by different algorithms for 2D epicenter estimation
 on a flat Earth
 """
 import sys
@@ -34,8 +34,8 @@ if len(src) > 1:
 
 log.info("Generating synthetic travel-time data")
 srcs, recs = ft.utils.connect_points(src, rec_points)
-ptime = ft.seis.ttime2d.straight(model, 'vp', srcs, recs)
-stime = ft.seis.ttime2d.straight(model, 'vs', srcs, recs)
+ptime = ft.seismic.ttime2d.straight(model, 'vp', srcs, recs)
+stime = ft.seismic.ttime2d.straight(model, 'vs', srcs, recs)
 error_level = 0.1
 ttr_true = stime - ptime
 ttr, error = ft.utils.contaminate(ttr_true, error_level, percent=True,
@@ -57,7 +57,7 @@ initial = initial[0]
 log.info("Will solve the inverse problem using Newton's method")
 nsolver = ft.inversion.gradient.newton(initial)
 newton = [initial]
-iterator = ft.seis.epic2d.homogeneous(ttr, recs, vp, vs, nsolver, iterate=True)
+iterator = ft.seismic.epic2d.homogeneous(ttr, recs, vp, vs, nsolver, iterate=True)
 for e, r in iterator:
     newton.append(e)
 newton_predicted = ttr - r
@@ -65,7 +65,7 @@ newton_predicted = ttr - r
 log.info("and the Steepest Descent method")
 sdsolver = ft.inversion.gradient.steepest(initial)
 steepest = [initial]
-iterator = ft.seis.epic2d.homogeneous(ttr, recs, vp, vs, sdsolver, iterate=True)
+iterator = ft.seismic.epic2d.homogeneous(ttr, recs, vp, vs, sdsolver, iterate=True)
 for e, r in iterator:
     steepest.append(e)
 steepest_predicted = ttr - r
@@ -73,7 +73,7 @@ steepest_predicted = ttr - r
 log.info("... and also the Levenberg-Marquardt algorithm for comparison")
 lmsolver = ft.inversion.gradient.levmarq(initial)
 levmarq = [initial]
-iterator = ft.seis.epic2d.homogeneous(ttr, recs, vp, vs, lmsolver, iterate=True)
+iterator = ft.seismic.epic2d.homogeneous(ttr, recs, vp, vs, lmsolver, iterate=True)
 for e, r in iterator:
     levmarq.append(e)
 levmarq_predicted = ttr - r
@@ -81,7 +81,7 @@ levmarq_predicted = ttr - r
 log.info("Build a map of the goal function")
 shape = (100, 100)
 xs, ys = ft.gridder.regular(area, shape)
-goals = ft.seis.epic2d.mapgoal(xs, ys, ttr, recs, vp, vs)
+goals = ft.seismic.epic2d.mapgoal(xs, ys, ttr, recs, vp, vs)
 
 log.info("Plotting")
 ft.vis.figure(figsize=(14,4))