+++ cookbook/seismic_profile_vertical.py
@@ -1,5 +1,5 @@
 """
-Seis: Invert vertical seismic profile (VSP) traveltimes for the velocity of the
+seismic: Invert vertical seismic profile (VSP) traveltimes for the velocity of the
 layers
 """
 import numpy
@@ -21,18 +21,18 @@ thickness, velocity = ft.vis.map.draw_layers(area, figure.gca())
 log.info("Generating synthetic data")
 zp = numpy.arange(zmin + 0.5, zmax, 0.5)
 tts, error = ft.utils.contaminate(
-    ft.seis.profile.vertical(thickness, velocity, zp),
+    ft.seismic.profile.vertical(thickness, velocity, zp),
     0.02, percent=True, return_stddev=True)
 
 log.info("Preparing for the inversion")
 damping = 100.
 estimates = []
 for i in xrange(30):
-    p, r = ft.seis.profile.ivertical(ft.utils.contaminate(tts, error), zp,
+    p, r = ft.seismic.profile.ivertical(ft.utils.contaminate(tts, error), zp,
         thickness, damping=damping)
     estimates.append(1./p)
 estimate = ft.utils.vecmean(estimates)
-predicted = ft.seis.profile.vertical(thickness, estimate, zp)
+predicted = ft.seismic.profile.vertical(thickness, estimate, zp)
 
 log.info("Plotting results...")
 ft.vis.figure(figsize=(12,5))