+++ cookbook/seismic_profile_vertical_interactive.py
@@ -1,5 +1,5 @@
 """
-Seis: Interactive forward modeling of 1D vertical seismic profile (VSP) data in
+seismic: Interactive forward modeling of 1D vertical seismic profile (VSP) data in
 layered media
 """
 import numpy