+++ cookbook/seismic_wavefd_elastic_sh_small.py
@@ -1,5 +1,5 @@
 """
-Seis: 2D finite difference simulation of elastic SH wave propagation
+seismic: 2D finite difference simulation of elastic SH wave propagation
 """
 import numpy as np
 from matplotlib import animation
@@ -8,7 +8,7 @@ import fatiando as ft
 log = ft.logger.get()
 
 # Make a wave source from a mexican hat wavelet
-sources = [ft.seis.wavefd.MexHatSource(25, 25, 100, 0.5, delay=1.5)]
+sources = [ft.seismic.wavefd.MexHatSource(25, 25, 100, 0.5, delay=1.5)]
 # Set the parameters of the finite difference grid
 shape = (50, 50)
 spacing = (1000, 1000)
@@ -21,7 +21,7 @@ svel = 3000*np.ones(shape)
 # computations take place at each iteration in the for loop bellow
 dt = 0.05
 maxit = 300
-timesteps = ft.seis.wavefd.elastic_sh(spacing, shape, svel, dens, dt, maxit,
+timesteps = ft.seismic.wavefd.elastic_sh(spacing, shape, svel, dens, dt, maxit,
     sources, padding=0.5)
 
 # This part makes an animation using matplotlibs animation API