+++ fatiando/__init__.py
@@ -6,7 +6,7 @@ Modules for each geophysical method are group in subpackages:
 
 * :mod:`pot <fatiando.pot>`:
   Potential fields
-* :mod:`seis <fatiando.seis>`:
+* :mod:`seismic <fatiando.seismic>`:
   Seismics and seismology
 * :mod:`geothermal <fatiando.geothermal>`:
   Geothermal heat transfer modeling
@@ -33,7 +33,7 @@ implementing inverse problems. There you'll find common regularizing functions,
 linear inverse problem solvers, and non-linear gradient solvers. This package
 is generaly only used from inside Fatiando itself, not when using Fatiando in
 scripts. For usage examples, see the source of modules
-:mod:`fatiando.seis.epic2d` and :mod:`fatiando.pot.basin2d`.
+:mod:`fatiando.seismic.epic2d` and :mod:`fatiando.pot.basin2d`.
 
 See the documentation for each module to find out more about what they do and
 how to use them.
@@ -50,7 +50,7 @@ from fatiando import logger
 from fatiando import gridder
 from fatiando import mesher
 from fatiando import pot
-from fatiando import seis
+from fatiando import seismic
 from fatiando import utils
 from fatiando import geothermal
 from fatiando import gui