+++ fatiando/seismic/__init__.py
@@ -1,0 +1,21 @@
+"""
+Various tools for seismic and seismology, like direct modeling, inversion
+(tomography), epicenter determination, etc.
+
+**Forward modeling and inversion**
+
+* :mod:`~fatiando.seismic.ttime2d`: 2D seismic ray travel-time modeling
+* :mod:`~fatiando.seismic.epic2d`: 2D epicenter determination
+* :mod:`~fatiando.seismic.profile`: Modeling and inversion of seismic profiling
+* :mod:`~fatiando.seismic.wavefd`: Finite difference solution of the 2D elastic
+  wave equation
+
+**Tomography**
+
+* :mod:`~fatiando.seismic.srtomo`: 2D straight-ray tomography
+
+----
+
+"""
+
+from fatiando.seismic import epic2d, ttime2d, srtomo, profile, wavefd