+++ doc/changelog.rst
@@ -43,8 +43,8 @@ Version 0.1
   methods for potential fields
 * Added module :ref:`fatiando.pot.euler <fatiando_pot_euler>` with Euler
   deconvolution methods for potential field data
-* Added module :ref:`fatiando.seis.wavefd <fatiando_seis_wavefd>` with 2D Finite
-  Difference simulations of elastic seismic waves
+* Added module :ref:`fatiando.seismic.wavefd <fatiando_seismic_wavefd>` with 2D
+  Finite Difference simulations of elastic seismic waves
 * Added unit conversion functions to :ref:`fatiando.utils <fatiando_utils>`
 * If building inplace or packaging, the setup script puts the Mercurial
   changeset hash in a file. Then :ref:`fatiando.logger.header <fatiando_logger>`